+++ metadata.rb
@@ -8,3 +8,7 @@ version           '0.0.1'
 source_url        'https://github.com/sous-chefs/vscode'
 issues_url        'https://github.com/sous-chefs/vscode/issues'
 chef_version '>= 13.0'
+
+%w(ubuntu debian redhat amazon centos windows fedora).each do |os|
+  supports os
+end